Use Unix socket
This sets the default configuration to listen on a Unix socket.
Since the backend on our reference instance is proxied by a local
web server, there is no need to for it to bind on a TCP listener.

We are using `bind 'unix:///run/paste/puma'` on our reference instance
(paste.opensuse.org) since some months already, this patch slightly
adjusts the configuration to allow for the path to be overwritten using
an environment variable.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>