Serve text files inline
Text-only pastes are preferably viewed in the browser instead of
downloaded when using their "raw" links.
To have browsers render such raw files natively instead of saving them
(which, depending on the browser and user defined browser preferences,
either comes in the form of a download to a pre-defined directory or of
a "Save as" dialogue), have the Content-Disposition header option get set
to "inline" instead of "attachment" for text/plain content.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>