Only clean up pastes which exist
If pastes would be deleted administratively before the scheduled expiry
cleanup took place, the cleanup job would, once the schedule hit,
return "Error performing PastesCleanupJob" along with a long traceback
about the failure.
Handle the situation gracefully and avoid the failure by validating whether
the paste exists before attempting to delete it.
To achieve this, the paste ID instead of the paste object must be passed
to the job, as otherwise the object in the queue might no longer be
pointing to any valid paste, not allowing for validation of it inside the
function.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>